Moved from seven static (dddd) colorN setting keys to one a(dddd)
this allows for a dynamic amount of color buttons within the toolbar
Additionally:
- Fixed issue regarding de-/ serialization of drawings that occured due to the change from float to double for the RGBA values of a stroke.
- Fixed issue with the navigation pane selecting the first instead of the active document.
- Replaced "^    " with "^\t"
- Maybe fixed compilation issues on Darwin (less likely MinGW too)

+++ drawing.h
@@ -11,7 +11,7 @@
 
 class CStroke {
 public:
-	double r,g,b,a;
+	float r,g,b,a;
 	std::vector<float> xcoords;
 	std::vector<float> ycoords;
 	std::vector<float> pcoords;
